app/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==21.4.0
backcall==0.2.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
fonttools==4.28.5
greenlet==1.1.2
htmlmin==0.1.12
idna==3.3
ImageHash==4.2.1
importlib-resources==5.4.0
ipykernel==6.6.1
ipython==7.31.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.0.1
jsonschema==4.3.3
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kaggle==1.5.12
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
missingno==0.5.0
mistune==0.8.4
multimethod==1.6
nbclient==0.5.9
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
notebook==6.4.6
numpy==1.22.0
packaging==21.3
pandas==1.3.5
pandas-profiling==3.1.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
phik==0.12.0
pickleshare==0.7.5
Pillow==9.0.0
#pkg_resources==0.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
ptyprocess==0.7.0
py4j==0.10.9.2
pycparser==2.21
pydantic==1.9.0
Pygments==2.11.1
pyparsing==3.0.6
pyrsistent==0.18.0
pyspark==3.2.0
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
requests==2.27.1
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
SQLAlchemy==1.4.29
sweetviz==2.1.3
tangled-up-in-unicode==0.1.0
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.7
visions==0.7.4
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
zipp==3.7.0
